[FLINK-16653][network][tests] Implement MockResultPartitionWriter base for simplifying tests
At the moment there are at-least four implementations of `ResultPartitionWriter` interface used in unit tests.
And there are about ten methods to be implemented for `ResultPartitionWriter` and most of them are dummy in tests.

When we want to extend the methods for `ResultPartitionWriter`, the above four dummy implementations in tests have to be adjusted as well, to waste a bit efforts.

Therefore the MockResultPartitionWriter is proposed to implement the basic dummy methods for `ResultPartitionWriter`,
and the previous four instances can all extend it to only implement one or two methods based on specific requirements in tests.
It will probably only need to adjust the MockResultPartitionWriter when extending the `ResultPartitionWriter` interface.

This closes #11441.
