+++ flink-runtime/src/test/java/org/apache/flink/runtime/io/network/api/writer/AbstractCollectingResultPartitionWriter.java
@@ -18,19 +18,16 @@
 
 package org.apache.flink.runtime.io.network.api.writer;
 
-import org.apache.flink.runtime.io.AvailabilityProvider;
 import org.apache.flink.runtime.io.network.buffer.Buffer;
 import org.apache.flink.runtime.io.network.buffer.BufferBuilder;
 import org.apache.flink.runtime.io.network.buffer.BufferConsumer;
 import org.apache.flink.runtime.io.network.buffer.BufferProvider;
-import org.apache.flink.runtime.io.network.partition.ResultPartitionID;
+import org.apache.flink.runtime.io.network.partition.MockResultPartitionWriter;
 
-import javax.annotation.Nullable;
 import javax.annotation.concurrent.ThreadSafe;
 
 import java.io.IOException;
 import java.util.ArrayDeque;
-import java.util.concurrent.CompletableFuture;
 
 import static org.apache.flink.util.Preconditions.checkNotNull;
 import static org.apache.flink.util.Preconditions.checkState;
@@ -39,31 +36,12 @@ import static org.apache.flink.util.Preconditions.checkState;
  * {@link ResultPartitionWriter} that collects output on the List.
  */
 @ThreadSafe
-public abstract class AbstractCollectingResultPartitionWriter implements ResultPartitionWriter {
+public abstract class AbstractCollectingResultPartitionWriter extends MockResultPartitionWriter {
 	private final BufferProvider bufferProvider;
 	private final ArrayDeque<BufferConsumer> bufferConsumers = new ArrayDeque<>();
 
 	public AbstractCollectingResultPartitionWriter(BufferProvider bufferProvider) {
 		this.bufferProvider = checkNotNull(bufferProvider);
-	}
-
-	@Override
-	public void setup() {
-	}
-
-	@Override
-	public ResultPartitionID getPartitionId() {
-		return new ResultPartitionID();
-	}
-
-	@Override
-	public int getNumberOfSubpartitions() {
-		return 1;
-	}
-
-	@Override
-	public int getNumTargetKeyGroups() {
-		return 1;
 	}
 
 	@Override
@@ -108,25 +86,6 @@ public abstract class AbstractCollectingResultPartitionWriter implements ResultP
 	@Override
 	public void flush(int subpartitionIndex) {
 		flushAll();
-	}
-
-	@Override
-	public void close() {
-	}
-
-	@Override
-	public void fail(@Nullable Throwable throwable) {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public void finish() {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public CompletableFuture<?> getAvailableFuture() {
-		return AvailabilityProvider.AVAILABLE;
 	}
 
 	protected abstract void deserializeBuffer(Buffer buffer) throws IOException;
